requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.42
botocore==1.24.42
Django==4.0.4
djangorestframework==3.13.1
jmespath==1.0.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
